Move Utf8String and Utf8Span code to src/libraries.
This is the first step in getting Utf8String and Utf8Span working on netstandard2.0.

Contributes to #29442
