+++ src/coreclr/src/System.Private.CoreLib/System.Private.CoreLib.csproj
@@ -299,23 +299,7 @@
     </Compile>
   </ItemGroup>
   <ItemGroup Condition="'$(FeatureUtf8String)' == 'true'">
-    <Compile Include="$(BclSourcesRoot)\System\Char8.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8Extensions.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8String.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8String.Comparison.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8String.Construction.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8String.Conversion.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8String.Enumeration.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8String.Manipulation.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8String.Searching.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Utf8StringSplitOptions.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Text\Utf8Span.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Text\Utf8Span.Comparison.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Text\Utf8Span.Conversion.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Text\Utf8Span.Enumeration.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Text\Utf8Span.Manipulation.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Text\Utf8Span.Searching.cs" />
-    <Compile Include="$(BclSourcesRoot)\System\Text\Utf8StringComparer.cs" />
+    <Compile Include="$(BclSourcesRoot)\System\Utf8String.CoreCLR.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="$(BclSourcesRoot)\System\Diagnostics\Eventing\XplatEventLogger.cs" Condition="'$(FeatureXplatEventSource)' == 'true'" />
