+++ src/libraries/System.Private.CoreLib/src/System/Char8.cs
