+++ src/libraries/System.Private.CoreLib/src/System/Text/Utf8Span.Enumeration.cs
