+++ src/libraries/System.Private.CoreLib/src/System/Utf8String.Manipulation.cs
